ecobuild-system/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4
python-multipart==0.0.6
pydantic==2.5.0
pymongo==4.6.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.0
email-validator==2.1.0
python-dotenv==1.0.0
